rename setting.svelte
