+++ src/components/setting-panel.svelte
